Here we go
On the 1st run, we turn each test case into a CNF formula (via JBMC),
do some reductions (via minisat) and store each clause as an element
of the set that will be minhashed later. We also store some annotations
from JBMC, telling us which SUT functions have been called by the TC
and where. This will disambiguate two tests that are identical but call
two different functions.

We need to a classpath and the directory where to run JBMC. This is
done by setting environment variables "FAST_JBMC_CP" and
"FAST_JBMC_DIR" (either at the cmd line or in an .env file).